+++ src/model/character/CharacterType.java
@@ -19,20 +19,20 @@ public enum CharacterType {
     double size;
     Set<Point> relativePointsUnder;
 
-    private double strength;
-    private double durability;
-    private double stamina;
-    private double eye;
-    private double arm;
-    private double agility;
-    private double knowledge;
-    private double focus;
-    private double charisma;
+    private int strength;
+    private int durability;
+    private int stamina;
+    private int eye;
+    private int arm;
+    private int agility;
+    private int knowledge;
+    private int focus;
+    private int charisma;
     private double dmgMin;
     private double dmgMax;
 
 
-    CharacterType(double size, double strength, double durability, double stamina, double eye, double arm, double agility, double knowledge, double focus, double charisma,
+    CharacterType(double size, int strength, int durability, int stamina, int eye, int arm, int agility, int knowledge, int focus, int charisma,
             double dmgMin, double dmgMax) {
         this.size = size;
         relativePointsUnder = MoveCalculator.calcRelativePointsUnder(this);
@@ -59,39 +59,39 @@ public enum CharacterType {
         return relativePointsUnder;
     }
 
-    public double getStrength() {
+    public int getStrength() {
         return strength;
     }
 
-    public double getDurability() {
+    public int getDurability() {
         return durability;
     }
 
-    public double getStamina() {
+    public int getStamina() {
         return stamina;
     }
 
-    public double getEye() {
+    public int getEye() {
         return eye;
     }
 
-    public double getArm() {
+    public int getArm() {
         return arm;
     }
 
-    public double getAgility() {
+    public int getAgility() {
         return agility;
     }
 
-    public double getKnowledge() {
+    public int getKnowledge() {
         return knowledge;
     }
 
-    public double getFocus() {
+    public int getFocus() {
         return focus;
     }
 
-    public double getCharisma() {
+    public int getCharisma() {
         return charisma;
     }
